// src/app/pages/models/GazeCevap.ts
export class GazeCastCevap 
{
    ID: number;

    GazeCastSoruID: number;
    soruAdi:string;
    Photo1Url: string;
    Photo1isTrue: boolean;
    dogruMu1: string;

    Photo2Url: string;
    Photo2isTrue: boolean;
    dogruMu2: string;
    Photo3Url: string;
    Photo3isTrue: boolean;
    dogruMu3: string;
    Photo4Url: string;
    Photo4isTrue: boolean;
    dogruMu4: string;

    Photo5Url: string;
    Photo5isTrue: boolean;
    dogruMu5: string;
    Photo6Url: string;
    Photo6isTrue: boolean;
    dogruMu6: string;
    Photo7Url: string;
    Photo7isTrue: boolean;
    dogruMu7: string;
    Photo8Url: string;
    Photo8isTrue: boolean;
    dogruMu8: string;

}